refactor(teams): split team.rs into focused submodules [OPE-429]
Split the 492-line team.rs into focused submodules:
- types.rs: data structures and enums
- validation.rs: field and workflow constraint validation
- recipe.rs: Goose Recipe conversion and orchestration instructions
- tests.rs: all unit tests
- mod.rs: public API, YAML parsing, and YamlDefinition trait impl
